Remove wrapper classes
